Changed data type of column 'amount' in subscriptions table from money to decimal for easier handling.

+++ server/sql/subscriptions/getSubscriptionsByUserId.sql
@@ -1,9 +1,9 @@
 SELECT 
   name, 
   nickname, 
-  to_char(due_date, 'YYYY-MM-DD') AS 'dueDate', 
-  reminder_days AS 'reminderDays', 
-  amount::numeric, 
-  subscription_uuid AS 'subscriptionUuid'
+  to_char(due_date, 'YYYY-MM-DD') AS "dueDate", 
+  reminder_days AS "reminderDays", 
+  amount, 
+  subscription_uuid AS "subscriptionUuid"
 FROM subscriptions 
 WHERE(user_id = ${userId});
